% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifH2ODeeplearning.R
\docType{data}
\name{LearnerClassifH2ODeeplearning}
\alias{LearnerClassifH2ODeeplearning}
\alias{mlr_learners_classif.h2odeeplearning}
\title{Classification H2O DeepLearning Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifH2ODeeplearning
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a deeplearning classification implemented in \code{\link[h2o:h2o.deeplearning]{h2o::h2o.deeplearning()}} in package \CRANpkg{h2o}.
}
\keyword{datasets}
