% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrH2ODeeplearning.R
\docType{data}
\name{LearnerRegrH2ODeeplearning}
\alias{LearnerRegrH2ODeeplearning}
\alias{mlr_learners_regr.h2odeeplearning}
\title{Regression H2O DeepLearning Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrH2ODeeplearning
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a deeplearning regression implemented in \code{\link[h2o:h2o.deeplearning]{h2o::h2o.deeplearning()}} in package \CRANpkg{h2o}.
}
\keyword{datasets}
