% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.h2o-package}
\alias{mlr3learners.h2o}
\alias{mlr3learners.h2o-package}
\title{mlr3learners.h2o: Learners from h2o package for mlr3}
\description{
Adds `h2o.deeplearning()` from the h2o package to
    mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
